feat: initial validation setup and major route segregation refactoring
