+++ src/main/scala/com/tudux/taxi/http/routes/ExtraInfoRoutes.scala
@@ -1,0 +1,61 @@
+package com.tudux.taxi.http.routes
+
+import akka.actor.{ActorRef, ActorSystem}
+import akka.http.scaladsl.marshallers.sprayjson.SprayJsonSupport
+import akka.http.scaladsl.model.StatusCodes
+import akka.http.scaladsl.server.Directives.{entity, _}
+import akka.http.scaladsl.server.Route
+import akka.pattern.ask
+import akka.util.Timeout
+import com.tudux.taxi.actors.extrainfo.TaxiTripExtraInfo
+import com.tudux.taxi.actors.extrainfo.TaxiTripExtraInfoCommand.GetTaxiTripExtraInfo
+import com.tudux.taxi.http.helpers.RouteFormatters._
+import com.tudux.taxi.http.helpers.RoutePayloads._
+import de.heikoseeberger.akkahttpcirce.FailFastCirceSupport._
+import io.circe.generic.auto._
+import spray.json._
+
+import scala.concurrent.ExecutionContext
+
+
+
+
+case class ExtraInfoRoutes(taxiTripActor: ActorRef)(implicit system: ActorSystem, dispatcher: ExecutionContext,timeout: Timeout ) extends SprayJsonSupport
+  with TaxiCostStatProtocol
+  with TaxiTimeInfoStatProtocol
+  with TaxiPassengerInfoProtocol
+  with TaxiExtraInfoProtocol
+  with CalculateDistanceCostProtocol
+  with CalculateAverageTripTimeProtocol
+  with GetAverageTipAmountProtocol
+  with GetTotalLoadedResponseProtocol
+{
+
+  val routes: Route = {
+      pathPrefix("api" / "yellowtaxi" / "extrainfo") {
+        get {
+          path(Segment) { tripId =>
+            println(s"Received some statID $tripId")
+            complete(
+              (taxiTripActor ? GetTaxiTripExtraInfo(tripId.toString))
+                //.mapTo[Option[TaxiExtraInfoStat]]
+                .mapTo[TaxiTripExtraInfo]
+                .map(_.toJson.prettyPrint)
+                .map(toHttpEntity)
+            )
+          }
+        } ~
+          put {
+            path(Segment) { tripId =>
+              put {
+                entity(as[UpdateExtraInfoRequest]) { request =>
+                  taxiTripActor ! request.toCommand(tripId)
+                  complete(StatusCodes.OK)
+                }
+              }
+            }
+          }
+      }
+  }
+}
+
