+++ src/main/scala/com/tudux/taxi/http/routes/Ping.scala
@@ -1,9 +1,9 @@
-package com.tudux.taxi.http
+package com.tudux.taxi.http.routes
 
 import akka.actor.ActorSystem
 import akka.http.scaladsl.model.HttpResponse
-import akka.http.scaladsl.model.StatusCodes._
-import akka.http.scaladsl.server.Directives._
+import akka.http.scaladsl.model.StatusCodes.OK
+import akka.http.scaladsl.server.Directives.{complete, get, path}
 
 case class Ping(system: ActorSystem) {
   val routes = path("ping") {
